+++ Beta Version/main.cpp
@@ -1,30 +1,22 @@
 #include "winsimple.hpp"
-
+#include "winsimple-controls.hpp"
 
 
 int main()
 {
-	ws::Rectangle rect(300,100);
-	rect.setPosition(3,3);
-	rect.setBorderWidth(3);
-	rect.setBorderColor(ws::Hue::blue);
-	rect.setFillColor(ws::Hue::yellow);
-	
-	ws::Hue hue = ws::Hue::yellow;
-	ws::Hue::HSV hsv = hue.toHSV();
-	
-	
 	ws::Window window(960,540,"");
-	window.enableChromaKey(ws::Hue::pink);
+	window.setIcon("logoMarkYellow.ico");
+	
+	
+	ws::Balloon(window,"Winsimple Says you have Windows " + ws::getWindowsVersion() + "!","Winsimple");
 	
 	while(window.isOpen())
 	{
-		hsv.h+=0.1;
-		rect.setBorderColor(hsv.toHue());
 		
-		window.clear(ws::Hue::pink);
-		window.draw(rect);
-		window.display();
+		
 	}
+
+
+	system("pause");
 	return 0;
 }